+++ app/Http/Controllers/to_do_list/ToDoPlansController.php
@@ -20,21 +20,32 @@ class ToDoPlansController extends Controller
 
     public function create(CreatePlanRequest $request, ToDoList $list)
     {
+        \Gate::authorize('list-owner', $list);
+
         return self::response($this->plansService->create($request->validated(), $list));
     }
 
     public function delete(ToDoPlan $plan)
     {
+        $this->authorize('actions', $plan);
+
         return self::response($this->plansService->delete($plan));
     }
 
     public function change(\Request $request, ToDoPlan $plan, ToDoList $newList = null)
     {
+        $this->authorize('actions', $plan);
+        if (isset($newList)) {
+            \Gate::authorize('list-owner', $newList);
+        }
+
         return self::response($this->plansService->change($request, $plan, $newList));
     }
 
     public function get(GetRequest $request, ToDoList $list)
     {
+        \Gate::authorize('list-owner', $list);
+
         return self::response([
             'message' => $this->plansService->plans($request->validated(), $list),
             'code' => 200
@@ -43,6 +54,8 @@ class ToDoPlansController extends Controller
 
     public function markComplete(ToDoPlan $plan)
     {
+        $this->authorize('actions', $plan);
+
         return self::response($this->plansService->change(['complete' => true], $plan));
     }
 }
